Conflicts:
	infogami/infobase/tests/test_client.py

+++ infogami/infobase/_dbstore/store.py
@@ -30,6 +30,8 @@ The following indexer allows querying for books using lowercase titles and books
             
 """
 
+from __future__ import with_statement
+
 import simplejson
 import web
 
@@ -58,41 +60,83 @@ class Store:
         self.listener and self.listener(name, data)
 
     def get_json(self, key):
-        row = self.get_row(key)
-        return row and row.json
+        d = self.get(key)
+        return d and simplejson.dumps(d)
     
     def get(self, key):
-        json = self.get_json(key)
-        return json and simplejson.loads(json)
+        row = self.get_row(key)
+        return row and self._get_doc(row)
+            
+    def _get_doc(self, row):
+        doc = simplejson.loads(row.json)
+        doc['_key'] = row.key
+        doc['_rev'] = str(row.id)
+        return doc
     
-    def put(self, key, data):
-        self.put_json(key, simplejson.dumps(data))
-    
-    def put_json(self, key, json):
+    def put(self, key, doc):
+        if doc.get("_delete") == True:
+            return self.delete(key, doc.get("_rev"))
+        
+        # conflict check is enabled by default. It can be disabled by passing _rev=None in the document.
+        if "_rev" in doc and doc["_rev"] is None:
+            enable_conflict_check = False
+        else:
+            enable_conflict_check = True
+            
+        doc.pop("_key", None)
+        rev = doc.pop("_rev", None)
+        
+        json = simplejson.dumps(doc)
+        
         tx = self.db.transaction()
         try:
             row = self.get_row(key, for_update=True)
             if row:
-                self.db.query("UPDATE store SET json=$json WHERE key=$key", vars=locals())
+                if enable_conflict_check and str(row.id) != str(rev):
+                    raise common.Conflict(key=key, message="Document update conflict")
+                
                 self.delete_index(row.id)
-                id = row.id
+                
+                # store query results are always order by id column.
+                # It is important to update the id so that the newly modified
+                # records show up first in the results.
+                self.db.query("UPDATE store SET json=$json, id=nextval('store_id_seq') WHERE key=$key", vars=locals())
+                
+                id = self.get_row(key=key).id
             else:
                 id = self.db.insert("store", key=key, json=json)
 
-            data = simplejson.loads(json)                
-            self.add_index(id, key, data)
+            self.add_index(id, key, doc)
+            
+            doc['_key'] = key
+            doc['_rev'] = str(id)
         except:
             tx.rollback()
             raise
         else:
             tx.commit()
-            self.fire_event("store.put", {"key": key, "data": data})
+            self.fire_event("store.put", {"key": key, "data": doc})
+            
+        return doc
+
+    def put_many(self, docs):
+        """Stores multiple docs in a single transaction.
+        """
+        with self.db.transaction():
+            for doc in docs:
+                key = doc['_key']
+                self.put(key, doc)
     
-    def delete(self, key):
+    def put_json(self, key, json):
+        self.put(key, simplejson.loads(json))
+    
+    def delete(self, key, rev=None):
         tx = self.db.transaction()
         try:
-            row = self.get_row(key)
+            row = self.get_row(key, for_update=True)
             if row:
+                if rev is not None and str(row.id) != str(rev):
+                    raise common.Conflict(key=key, message="Document update conflict")
                 self.delete_row(row.id)
         except:
             tx.rollback()
@@ -106,7 +150,7 @@ class Store:
         self.db.delete("store_index", where="store_id=$id", vars=locals())
         self.db.delete("store", where="id=$id", vars=locals())
         
-    def query(self, type, name, value, limit=100, offset=0):
+    def query(self, type, name, value, limit=100, offset=0, include_docs=False):
         """Query the json store.
         
         Returns keys of all documents of the given type which have (name, value) in the index.
@@ -114,7 +158,7 @@ class Store:
         All the documents are returned when the type is None.
         """
         if type is None:
-            rows = self.db.select("store", what="key", limit=limit, offset=offset, order="store.id desc", vars=locals())
+            rows = self.db.select("store", what="store.*", limit=limit, offset=offset, order="store.id desc", vars=locals())
         else:
             tables = ["store", "store_index"]
             wheres = ["store.id = store_index.store_id", "type = $type"]
@@ -123,9 +167,15 @@ class Store:
                 wheres.append("name='_key'")
             else:
                 wheres.append("name=$name AND value=$value")
-            rows = self.db.select(tables, what='store.key', where=" AND ".join(wheres), limit=limit, offset=offset, order="store.id desc", vars=locals())
+            rows = self.db.select(tables, what='store.*', where=" AND ".join(wheres), limit=limit, offset=offset, order="store.id desc", vars=locals())
             
-        return [{"key": r.key} for r in rows]
+        def process_row(row):
+            if include_docs:
+                return {"key": row.key, "doc": self._get_doc(row)}
+            else:
+                return {"key": row.key}
+                
+        return [process_row(row) for row in rows]
     
     def delete_index(self, id):
         self.db.delete("store_index", where="store_id=$id", vars=locals())
@@ -139,6 +189,8 @@ class Store:
         ignored = ["type"]
         for name, value in set(self.indexer.index(data)):
             if not name.startswith("_") and name not in ignored:
+                if isinstance(value, bool):
+                    value = str(value).lower()
                 d.append(web.storage(store_id=id, type=type, name=name, value=value))
         if d:
             self.db.multiple_insert('store_index', d)
